+++ Source/WebSocket/Android/AndroidWebSocketProvider.h
